add exceptions

+++ teams/src/storage/group/groupCreate.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { AppError } from '@utils/AppError';
 
 import { groupsGetAll } from './groupsGetAll';
 import { GROUP_COLLECTION } from '@storage/storageConfig';
@@ -7,6 +8,12 @@ export async function groupCreate(newGroup: string) {
   try {
 
     const storedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if(groupAlreadyExists) {
+      throw new AppError('There is already a group registered with that name.')
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup])
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
